refactor: optimize e2e tests
