+++ .github/workflows/continuous-integration-e2e.yaml
@@ -50,7 +50,7 @@ jobs:
 
       - name: 🔨 Build
         run: |
-          yarn install --immutable --inline-builds
+          yarn install --immutable --inline-builds --mode=skip-build
           yarn build
           docker build --no-cache .
         env:
@@ -67,18 +67,24 @@ jobs:
           OGMIOS_URL: 'ws://ogmios:1340'
           POSTGRES_PORT: '5435'
 
-      - name: Wait for some epochs
+      - name: Wait for network init
         run: |
-          yarn workspace @cardano-sdk/e2e wait-for-network
-        env:
-          DB_SYNC_CONNECTION_STRING: 'postgresql://postgres:doNoUseThisSecret!@localhost:5435/cexplorer'
+          yarn workspace @cardano-sdk/e2e wait-for-network-init
 
-      - name: 🔬 Test - e2e - wallet
+      - name: 🔬 Test - e2e - wallet at epoch 0
         run: |
-          yarn workspace @cardano-sdk/e2e test:wallet
-          yarn workspace @cardano-sdk/e2e test:providers
+          yarn workspace @cardano-sdk/e2e test:wallet:epoch0
           yarn workspace @cardano-sdk/e2e test:projection
+
+      - name: Wait for epoch 3
+        run: |
+          yarn workspace @cardano-sdk/e2e wait-for-network-epoch-3
+
+      - name: 🔬 Test - e2e - wallet at epoch 3
+        run: |
+          yarn workspace @cardano-sdk/e2e test:wallet:epoch3
           yarn workspace @cardano-sdk/e2e test:pg-boss
+          yarn workspace @cardano-sdk/e2e test:providers
         env:
           STAKE_POOL_PROVIDER_URL: 'http://localhost:4000/'
