+++ packages/cardano-services/docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.9'
-
 x-logging: &logging
   logging:
     driver: 'json-file'
