+++ packages/e2e/docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.9'
-
 x-logging: &logging
   logging:
     driver: 'json-file'
@@ -68,9 +66,6 @@ services:
       - ./local-network/config/network:/config
 
   cardano-smash:
-    depends_on:
-      cardano-db-sync:
-        condition: service_healthy
     volumes:
       - ./local-network/config/network:/config
 
@@ -113,6 +108,7 @@ services:
   stake-pool-projector:
     environment:
       METADATA_JOB_RETRY_DELAY: 60
+      POOLS_METRICS_INTERVAL: 50
 
 volumes:
   sdk-ipc:
