+++ packages/e2e/test/wallet_epoch_0/PersonalWallet/metadata.test.ts
