+++ packages/e2e/test/wallet_epoch_0/PersonalWallet/mint.test.ts
